obtener/buscar/actualizar presupuesto
obtener/buscar/actualizar presupuesto

+++ src/main/java/com/gestion/SNYA/repositorio/InstitucionRepositorio.java
@@ -7,7 +7,7 @@ import com.gestion.SNYA.modelo.Institucion;
 
 @Repository
 
-public interface InstitucionRepositorio extends JpaRepository<Institucion, Long> {
+public interface InstitucionRepositorio extends JpaRepository<Institucion,Integer> {
 	
 
 }
