+++ src/main/java/com/gestion/SNYA/repositorio/PresupuestoRepositorio.java
@@ -1,8 +1,11 @@
 package com.gestion.SNYA.repositorio;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import com.gestion.SNYA.modelo.Presupuesto;
+
 @Repository
-public interface PresupuestoRepositorio extends JpaRepository<Presupuesto, Long>{
+public interface PresupuestoRepositorio extends JpaRepository<Presupuesto, Integer>{
     
 }
